# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API softskills",
        "description": "Documentation API softskills",
        "contact": {
            "name": "Kouame Banh",
            "email": "kouamebanh@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/api/v2",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v2/article/{articleId}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get list of comments",
                "description": "Returns list of comments",
                "operationId": "getComments",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Comments"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comments not found"
                    },
                    "500": {
                        "description": "Error retrieving comments"
                    }
                }
            }
        },
        "/api/v2/add/comment/article/{articleId}": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create new comment",
                "description": "Returns comment data",
                "operationId": "storeComment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Comments"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comments"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    },
                    "500": {
                        "description": "Error creating comment"
                    }
                }
            }
        },
        "/api/v2/reply/comment/{commentId}/article/{articleId}": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Reply to a comment",
                "description": "Returns comment data",
                "operationId": "replyComment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Comments"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comments"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    },
                    "500": {
                        "description": "Error creating comment"
                    }
                }
            }
        },
        "/api/v2/replies/article/{articleId}/comment/{commentId}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get list of replies",
                "description": "Returns list of replies",
                "operationId": "getReplies",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Comments"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Replies not found"
                    },
                    "500": {
                        "description": "Error retrieving replies"
                    }
                }
            }
        },
        "/api/v2/comment/{id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get comment details",
                "description": "Returns comment details",
                "operationId": "getCommentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comments"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    },
                    "500": {
                        "description": "Error retrieving comment"
                    }
                }
            }
        },
        "/api/v2/update/comment/{id}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Update existing comment",
                "description": "Returns updated comment data",
                "operationId": "updateComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Comments"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comments"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    },
                    "500": {
                        "description": "Error updating comment"
                    }
                }
            }
        },
        "/api/v2/delete/comment/{id}": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete existing comment",
                "description": "Deletes a comment",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment deleted successfully"
                    },
                    "404": {
                        "description": "Comment not found"
                    },
                    "500": {
                        "description": "Error deleting comment"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Comments": {
                "title": "Comment",
                "description": "Modèle de commentaire",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "content": {
                        "type": "string",
                        "example": "Commentaire"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "article_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "user": {
                        "description": "User",
                        "type": "object",
                        "example": {
                            "id": 1,
                            "name": "John Doe",
                            "email": "QHn4o@example.com",
                            "email_verified_at": "2023-01-01T00:00:00Z",
                            "created_at": "2023-01-01T00:00:00Z",
                            "updated_at": "2023-01-01T00:00:00Z"
                        }
                    },
                    "article": {
                        "description": "Article",
                        "type": "object",
                        "example": {
                            "id": 1,
                            "title": "Article title",
                            "content": "Article content",
                            "user_id": 1,
                            "created_at": "2023-01-01T00:00:00Z",
                            "updated_at": "2023-01-01T00:00:00Z"
                        }
                    },
                    "replies": {
                        "description": "Replies",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comments"
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Utilisez un jeton JWT pour l'authentification.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}